refactor: scanKeys 함수 유틸로 공통 분리 (#265)
