+++ apps/api/src/room/room.service.ts
@@ -12,6 +12,7 @@ import { ProblemService } from '@/problem/problem.service';
 
 import { BattleService } from '../battle/battle.service';
 import { REDIS_CLIENT } from '../redis/redis.module';
+import { scanKeys } from '../redis/redis.util';
 import { RedisKeys } from '../redis/redis-key.constant';
 import { REDIS_TTL } from '../redis/redis-ttl.constant';
 
@@ -189,20 +190,9 @@ export class RoomService {
     await pipeline.exec();
   }
 
-  private async scanKeys(pattern: string): Promise<string[]> {
-    const keys: string[] = [];
-    const stream = this.redis.scanStream({ match: pattern, count: 100 });
-
-    return new Promise((resolve, reject) => {
-      stream.on('data', (batch: string[]) => keys.push(...batch));
-      stream.on('end', () => resolve(keys));
-      stream.on('error', reject);
-    });
-  }
-
   async listRooms(): Promise<Room[]> {
     // 방 정보를 담고 있는 모든 키 조회 (SCAN 사용으로 비블로킹)
-    const keys = await this.scanKeys(RedisKeys.room('*'));
+    const keys = await scanKeys(this.redis, RedisKeys.room('*'));
     if (!keys.length) return [];
 
     // 파이프라인으로 한 번에 조회 후 파싱
